Move ToonGlobals to core
