+++ core/ToonGlobals.py
